% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckeck_config.R
\name{check_cruncher_configuration}
\alias{check_cruncher_configuration}
\title{Check Cruncher configuration
This function checks if the Cruncher configuration is valid by verifying the presence of the \code{jwsacruncher} executable in the specified \code{cruncher_bin_directory}.}
\usage{
check_cruncher_configuration(
  cruncher_bin_directory = getOption("cruncher_bin_directory"),
  print_message = TRUE
)
}
\arguments{
\item{cruncher_bin_directory}{Path to the "bin" directory of the 'JWSACruncher'. By default
the value of the option \code{"cruncher_bin_directory"} is used.}

\item{print_message}{Logical.
If \code{TRUE}, a message will be printed indicating whether the configuration is valid or not.}
}
\description{
Check Cruncher configuration
This function checks if the Cruncher configuration is valid by verifying the presence of the \code{jwsacruncher} executable in the specified \code{cruncher_bin_directory}.
}
