% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cruncher.R
\encoding{UTF-8}
\name{cruncher}
\alias{cruncher}
\title{Lanceur de cruncher}
\usage{
cruncher(workspace,
  cruncher_bin_directory = getOption("cruncher_bin_directory"),
  param_file_path, log_file)
}
\arguments{
\item{workspace}{path to the workspace. By default a dialog box opened to choose the workspace.}

\item{cruncher_bin_directory}{path to the "bin" directory of the 'JWSACruncher'. The default
the value of the option \code{"cruncher_bin_directory"} is used.}

\item{param_file_path}{path to the parameter file of the cruncher. By default a .params file is search in the save directory of the workspace.}

\item{log_file}{nom du fichier qui contiendra la log du cruncher. Par défaut la log n'est pas exportée}
}
\value{
L'adresse du workspace.
}
\description{
Fonction qui permet de lancer le cruncher sur un workspace à partir d'un fichier de paramètres.
}
