% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_file.R
\name{default_param_file}
\alias{default_param_file}
\title{Get default parameters of the 'JWSACruncher'}
\usage{
default_param_file(v3 = getOption("is_cruncher_v3"))
}
\arguments{
\item{v3}{Boolean indicating if the parameters are the from a version 3.0.0 and higher of 'JWSACRuncher' (\code{v3 = TRUE}) or a lower version (\code{v3 = FALSE}). By default
the value of the option \code{"is_cruncher_v3"} is used (equals to \code{FALSE} by default).}
}
\description{
Get default parameters of the 'JWSACruncher'
}
\seealso{
\code{\link[=create_param_file]{create_param_file()}}, \code{\link[=read_param_file]{read_param_file()}}, \code{\link[=list2param_file]{list2param_file()}}, \code{\link[=cruncher_and_param]{cruncher_and_param()}}.
}
