% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_file.R
\name{list2param}
\alias{list2param}
\title{Create parameter file from list}
\usage{
list2param(dir_file_param, x)
}
\arguments{
\item{dir_file_param}{Path to the directory that will contains the parameter file "parameters.param".}

\item{x}{a list, for example create by \code{\link{read_param_file}} or \code{\link{default_param_file}}.}
}
\description{
Create parameter file from list
}
