% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cruncher.R
\encoding{UTF-8}
\name{multiprocessing_names}
\alias{multiprocessing_names}
\title{Get the names of the multiprocessings of a workspace}
\usage{
multiprocessing_names(workspace)
}
\arguments{
\item{workspace}{path to the workspace. By default a dialog box opened to choose the workspace.}
}
\value{
Un \code{data.frame} contenant le nom des multiprocessings sous JDemetra+ (colonne \code{name}) et
le nom des fichiers XML associés (colonne \code{file})
}
\description{
Fonction qui permet d'extraire le nom des multiprocessings sous JDemetra+ et les noms des fichiers XML associés.
}
