% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_file.R
\name{read_param_file}
\alias{read_param_file}
\title{Create parameter file for the 'JWSACruncher'}
\usage{
read_param_file(file)
}
\arguments{
\item{file}{Path to the parameter file.}
}
\description{
Create parameter file for the 'JWSACruncher'
}
