% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cruncher.R
\encoding{UTF-8}
\name{update_workspace}
\alias{update_workspace}
\title{Update a workspace}
\usage{
update_workspace(workspace = NULL, policy = "parameters",
  cruncher_bin_directory = getOption("cruncher_bin_directory"))
}
\arguments{
\item{workspace}{path to the workspace. By default a dialog box opened to choose the workspace.}

\item{policy}{méthode de rafraîchissement utilisée. Par défaut \code{policy = "parameters"} (paramètres re-estimés).
Les autres méthodes possibles sont :
\code{"outliers"} (les outliers sont identifiés et les paramètres re-estimés) ;
\code{"lastoutliers"} (les outliers sont ré-identifiés sur la dernière année et les paramètres re-estimés) ;
\code{"stochastic"} (le modèle arima et les outliers sont identifiés et les paramètres re-estimés) ;
\code{"complete"} (le modèle est complétement ré-estimé).}

\item{cruncher_bin_directory}{répertoire contenant contenant le dossier "bin" du cruncher.}
}
\value{
L'adresse du workspace.
}
\description{
Function to update a workspace without exporting the results.
}
