% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cruncher.R
\encoding{UTF-8}
\name{cruncher}
\alias{cruncher}
\title{Run the 'JWSACruncher'}
\usage{
cruncher(workspace,
  cruncher_bin_directory = getOption("cruncher_bin_directory"),
  param_file_path, log_file)
}
\arguments{
\item{workspace}{path to the workspace. By default a dialog box opens to choose the workspace.}

\item{cruncher_bin_directory}{path to the "bin" directory of the 'JWSACruncher'. By default default
the value of the option \code{"cruncher_bin_directory"} is used.}

\item{param_file_path}{path to the parameter file of the 'JWSACruncher' By default a .params file is search in the save directory of the workspace.}

\item{log_file}{name of the log file of 'JWSACruncher'. By default the log isn't exported.}
}
\value{
The path to the workspace.
}
\description{
Function to run the 'JWSACruncher' on a workspace from a parameter file.
}
\seealso{
\code{\link{cruncher_and_param}}, \code{\link{update_workspace}}.
}
