% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cruncher.R
\encoding{UTF-8}
\name{multiprocessing_names}
\alias{multiprocessing_names}
\title{Get the names of the multiprocessings of a workspace}
\usage{
multiprocessing_names(workspace)
}
\arguments{
\item{workspace}{Path to the workspace. By default a dialog box opens to choose the workspace.}
}
\value{
A \code{data.frame} containing the name of the multiprocessings that appears on 'JDemetra+' (column \code{name}) and
the name of the associated XML files (column \code{file}).
}
\description{
Function to get the name of the multiprocessings that appears on 'JDemetra+' and the name of the corresponding XML file.
}
\seealso{
\code{\link[=cruncher_and_param]{cruncher_and_param()}}.
}
